split exchange tests, cleanup names and includes
